idaes.accel dispatch layer, parity harness, and the first accelerated function
Adds the runtime bridge between idaes-pse and the optional idaes-accel wheel,
plus the parity machinery that keeps the two implementations honest, and wires
the first call site end to end.

idaes/accel/ (new package; `packages.find` already discovers it, so pyproject
needs no change beyond the optional `accel` extra)
- _loader: probes for the extension once, lazily, behind a lock. Never imports
  idaes at module scope (idaes/__init__.py builds the config block, so a reverse
  edge is a cycle) and never imports the extension at import time -- making its
  native runtime unconditional on `import idaes` would regress every user who
  has no acceleration installed. IDAES_ACCEL=auto|force|off; a misspelled value
  raises rather than silently meaning `auto`, so a CI job cannot believe it
  exercised a path it did not. IDAES_ACCEL_DISABLE is a per-key kill switch.
  Free-threaded CPython is detected and declined: abi3 wheels cannot load there.
- compat: the contract is one integer, compared for EXACT equality. Breakage is
  bidirectional, and `packaging` is not a dependency of idaes-pse, so version
  strings stay diagnostic-only.
- _registry: `@accelerate` decorates the pure-Python implementation and keeps it
  reachable; both impls stay available, which is what lets one parity test cover
  every function. Caches (generation, callable) as a single tuple so the hot path
  is one load and one compare and can never be seen half-updated. Preserves
  __signature__, without which Sphinx renders every wrapped function as
  (*args, **kwargs). Refuses to wrap classes: that would break issubclass() and
  the getattr-by-name lookups in parameter_sweep.
- parity/_discovery: cases are registered next to the implementation and found
  through an explicit call-site list, not package walking.

Parity harness
- One parametrized file covers every registered case; adding an accelerated
  function adds no test code. Tests call both implementations directly and never
  read IDAES_ACCEL, which is what lets the identical file run under `off` and
  `force` -- the env var is a CI concern.
- Split into three test functions by tier because idaes/conftest.py requires
  exactly one of unit/component/integration/performance, and a function-level
  marker plus a param-level marker would count as two.
- Bit-exact by default; a loosened tolerance requires a written justification,
  enforced in parity.register(). Also asserts inputs survive the call unchanged,
  since a Rust impl taking a mutable view of a NumPy buffer diverges silently.
- Optional deps gated per case with skipif rather than through Importorskipper,
  which skips whole modules and would drop every case for one missing import.

First call site: pysmo.sampling.prime_number_generator
- Extracted to a module-level function; the method delegates and keeps its
  signature and docstring. `accelerate` must never wrap a bound method or `self`
  crosses the boundary as the first positional argument.

BUG CAUGHT BY THE FULL-SUITE RUN, not by parity unit tests
- The Python function is untyped and its own suite calls it with a float:
  test_prime_number_generator_05 passes 2.9 and expects [2, 3, 5], because
  `len(prime_list) < 2.9` goes false at three. The i64 Rust signature raised
  TypeError where Python returned a list -- two failures above baseline under
  IDAES_ACCEL=force. The kernel now takes f64 and mirrors Python's comparison,
  including NaN and negative n terminating immediately. Float parity cases are
  registered so this cannot regress.
- Porting an untyped function means porting its coercion behaviour, not just its
  algorithm.

Environment: uv, pinned to CPython 3.14.7
- .python-version is now tracked (un-ignored from the upstream pyenv entry); it
  is a toolchain pin, not a per-developer scratch file.
- justfile gains dev-env / dev-env-accel / env-info / py-test-matrix and defaults
  wheel verification to the pinned interpreter. 3.10-3.13 are a compatibility
  matrix, not the default to develop against.

Config
- pylintrc: idaes_accel._core added to extension-pkg-allow-list (pylint refuses
  to introspect a C extension otherwise) and idaes_accel to ignored-modules,
  since its absence is a supported runtime state, not an import error.
- pyproject [tool.black]: exclude docs-code-update/.codex/.agents. CI runs
  `black --check .` over the whole root, and that vendored tree is not IDAES
  source. Same reasoning as the existing typos exclusion.

Verified on CPython 3.14.7 and 3.12
- accel suite: 12 passed / 13 skipped with no extension; 25 passed under force.
- pysmo suite: 34 failed / 898 passed under BOTH off and force -- identical to
  pristine upstream with these changes stashed, so zero regressions. Those 34
  are scratch-environment dependency drift, not this change.
- One cp310-abi3 wheel built under 3.12 loads and runs on 3.14.7.
- black 26.3.1 (the pinned version) clean over 901 files; typos clean;
  idaes/tests/test_headers.py passes; pylint 10.00/10 on the new code;
  just ci-fast and cargo deny/audit clean.

+++ rust/py/idaes-accel/src/lib.rs
@@ -29,9 +29,13 @@ const ABI_VERSION: u32 = 1;
 ///
 /// Returns a Python `list[int]`, not a NumPy array: `HaltonSampling.sample_points`
 /// indexes this and feeds the element into integer arithmetic.
+///
+/// `n` is `f64` because the Python function is untyped and is called with floats
+/// in the IDAES test suite. Accepting `i64` here would raise `TypeError` where
+/// the pure-Python path returns a list -- a divergence, not a stricter contract.
 #[pyfunction]
 #[pyo3(name = "pysmo__sampling__prime_number_generator")]
-fn pysmo_sampling_prime_number_generator(py: Python<'_>, n: i64) -> Vec<u64> {
+fn pysmo_sampling_prime_number_generator(py: Python<'_>, n: f64) -> Vec<u64> {
     // Pure Rust, no Python objects touched: safe to drop the GIL.
     py.detach(|| idaes_accel_core::sampling::prime_number_generator(n))
 }
